+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - Fixed handling of root incompatibilities.
 - Fixed an error where packages from custom repositories would not be found.
 - Fixed wildcard Python requirement being wrongly set in distributions metadata.
+- Fixed installation of packages from a custom repository.
 
 
 ## [0.10.3] - 2018-06-04